# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "ip_service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/asn": {
            "get": {
                "description": "ASN for the given IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "get ASN for the given IP",
                "operationId": "asn",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city": {
            "get": {
                "description": "city for the given IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "get city for the given IP",
                "operationId": "city",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collision": {
            "post": {
                "description": "takes two CIDR blocks and checks if they overlap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "Check if two CIDR blocks overlap",
                "operationId": "collision",
                "parameters": [
                    {
                        "description": " ",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiv1.CollisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/apiv1.CollisionReply"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coordinates": {
            "get": {
                "description": "coordinates for the given IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "get Coordinates for the given IP",
                "operationId": "coordinates",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/country": {
            "get": {
                "description": "country for the given IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "get country for the given IP",
                "operationId": "country",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/country-iso": {
            "get": {
                "description": "country ISO for the given IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "get Country ISO for the given IP",
                "operationId": "countryISO",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/lookup/{ip}": {
            "get": {
                "description": "takes query parameter ip and returns all information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "Look up all information for the given IP",
                "operationId": "lookUpIP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/model.ReplyLookUp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/whois/{ip}": {
            "get": {
                "description": "takes query parameter ip and returns whois information in JSON format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_service"
                ],
                "summary": "Whois information in JSON format",
                "operationId": "whois",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/rpsl.ASN"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiv1.CollisionReply": {
            "type": "object",
            "properties": {
                "collision": {
                    "type": "boolean"
                }
            }
        },
        "apiv1.CollisionRequest": {
            "type": "object",
            "required": [
                "ip_1",
                "ip_2"
            ],
            "properties": {
                "ip_1": {
                    "type": "string"
                },
                "ip_2": {
                    "type": "string"
                }
            }
        },
        "helpers.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "title": {
                    "type": "string"
                }
            }
        },
        "helpers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/helpers.Error"
                }
            }
        },
...
        "model.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "model.ReplyLookUp": {
            "type": "object",
            "properties": {
                "asn": {
                    "type": "integer"
                },
                "asn_organization": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "continent": {
                    "type": "string"
                },
                "coordinates": {
                    "$ref": "#/definitions/model.Coordinates"
                },
                "country": {
                    "type": "string"
                },
                "country_iso": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "ip_decimal": {
                    "type": "string"
                },
                "is_1918_network": {
                    "type": "boolean"
                },
                "is_eu": {
                    "type": "boolean"
                },
                "postal_code": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "region_code": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "whois": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/rpsl.Object"
                    }
                }
            }
        },
...
        "rpsl.ASN": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/rpsl.Object"
            }
        },
        "rpsl.Object": {
            "type": "object",
            "properties": {
                "abuse": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "abuse-c": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "abuse-mailbox": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "address": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "admin-c": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aggr-bndry": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aggr-mtd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "alias": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "as-name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "as-set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aut-num": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "auth": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "certif": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "changed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "components": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "correo": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dba": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "default": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "descr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "e-mail": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "export": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "export-comps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "export-via": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fax-no": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter-set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fingerpr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "geoidx": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "holes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ifaddr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "import": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "import-via": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inet-rtr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inet6num": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inetnum": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "interface": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ips": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "key-cert": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "last-modified": {
                    "type": "string"
                },
                "local-as": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mailto": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mbrs-by-ref": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "member-of": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mnt-by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mnt-lower": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mnt-nfy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mnt-routes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mntner": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-default": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-export": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-import": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-peer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mp-peering": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "netname": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "network": {
                    "type": "string"
                },
                "nic-hdl": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "notify": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "org": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "org-name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "org-type": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "origin": {
                    "type": "string"
                },
                "owner": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "owner-c": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ownerid": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "peer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "peering": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "peering-set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "person": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ping-hdl": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pingable": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "remarks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "responsible": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roa-uri": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "route-set": {
                    "description": "Route        []string `json:\"route,omitempty\" bson:\"route,omitempty\"`\nRoute6       []string `json:\"route6,omitempty\" bson:\"route6,omitempty\"`",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rs-in": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rs-out": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rtr-set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "support": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tech-c": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tel": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trouble": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "upd-to": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "website": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}